backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
pydantic==2.6.1
loguru==0.7.2
python-dotenv==1.0.1
pytest==8.0.0
python-multipart==0.0.9

# Document Parsing for Ingestor
PyMuPDF
pandas==2.2.0

# AI/LLM for Research Agent & CAM Gen
langchain==0.1.7
langchain-openai==0.0.6
langchain-google-genai>=1.0.0
openai==1.12.0
tavily-python==0.3.2

# Scoring Engine & ML Models
numpy
scikit-learn
joblib
psycopg2-binary
networkx
matplotlib

# Advanced Document AI & GNN Detection
torch
torch_geometric
transformers
torchvision
easyocr
opencv-python-headless
